rename litentry to heima
