+++ packages/apps-config/src/ui/logos/nodes/index.ts
@@ -59,6 +59,8 @@ export { nodesGeekSVG } from './generated/geekSVG.js';
 export { nodesGiantPNG } from './generated/giantPNG.js';
 export { nodesHanyonycashPNG } from './generated/hanyonycashPNG.js';
 export { nodesHashedPNG } from './generated/hashedPNG.js';
+export { nodesHeimaPaseoPNG } from './generated/heima-paseoPNG.js';
+export { nodesHeimaSVG } from './generated/heimaSVG.js';
 export { nodesHelixstreetPNG } from './generated/helixstreetPNG.js';
 export { nodesHumanodePNG } from './generated/humanodePNG.js';
 export { nodesHyperbridgePNG } from './generated/hyperbridgePNG.js';
@@ -88,9 +90,6 @@ export { nodesKulupuSVG } from './generated/kulupuSVG.js';
 export { nodesKusariSVG } from './generated/kusariSVG.js';
 export { nodesKylinPNG } from './generated/kylinPNG.js';
 export { nodesLaminarCircleSVG } from './generated/laminar-circleSVG.js';
-export { nodesLitentryPaseoSVG } from './generated/litentry-paseoSVG.js';
-export { nodesLitentryRococoPNG } from './generated/litentry-rococoPNG.js';
-export { nodesLitentryPNG } from './generated/litentryPNG.js';
 export { nodesLoomNetworkPNG } from './generated/loom_networkPNG.js';
 export { nodesMagnetPNG } from './generated/magnetPNG.js';
 export { nodesMandalaPNG } from './generated/mandalaPNG.js';
